deep-search/pyproject.toml
[project]
name = "deep-search"
version = "0.1.0"
description = "A production-ready fullstack research agent that uses Gemini to strategize, research, and synthesize comprehensive reports with human-in-the-loop collaboration"
authors = [
    {name = "Amit Dutta", email = "amitdutta@google.com"},
    {name = "Deepak Moonat", email = "dmoonat@google.com"},
    {name = "Elia Secchi", email = "eliasecchi@google.com"},
    {name = "Erwin Huizenga", email = "erwinh@google.com"},
    {name = "Lavi Nigam", email = "lavinigam@google.com"},
]
license = {text = "Apache License 2.0"}
readme = "README.md"
requires-python = ">=3.10,<3.13"

dependencies = [
    "google-adk>=1.8.0",
    "mcp>=0.1.0",
    "openai>=1.1.0",
    "anthropic>=0.1.0",
    "litellm>=1.1.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "pypdf>=4.0.0",
    "python-docx>=1.0.0",
    "weasyprint>=62.0",
    "markdown>=3.6",
    "websockets>=12.0",
    "apscheduler>=3.10.0",
    "matplotlib>=3.8.0",
    "plotly>=5.18.0",
    "fastapi>=0.110.0",
    "python-multipart>=0.0.9",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "nest-asyncio>=1.6.0",
    "agent-starter-pack>=0.14.1",
    "jupyter>=1.1.1",
    "ruff>=0.4.6",
    "mypy>=1.19.0",
    "codespell>=2.4.1",
    "types-pyyaml>=6.0.12.20240917",
    "types-requests>=2.32.4.20250913",
]

[tool.ruff]
# Point back to the Master template in the root
extend = "../../../pyproject.toml"

[tool.ruff.lint.isort]
# Treat both app and frontend as internal packages
known-first-party = ["app", "frontend", "deep_search"]

[tool.mypy]
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
check_untyped_defs = true
disallow_subclassing_any = true
warn_incomplete_stub = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
follow_imports = "silent"
ignore_missing_imports = true
explicit_package_bases = true
disable_error_code = ["misc", "no-untyped-call", "no-any-return"]
exclude = [".venv"]

[tool.hatch.build.targets.wheel]
packages = ["app", "frontend"]

[tool.pytest.ini_options]
pythonpath = "."
asyncio_default_fixture_loop_scope = "function"

[tool.agent-starter-pack]
base_template = "adk_base"
description = "A production-ready fullstack research agent that uses Gemini to strategize, research, and synthesize comprehensive reports with human-in-the-loop collaboration"
example_question = "A report on the latest Google I/O event"

[tool.agent-starter-pack.settings]
deployment_targets = ["agent_engine", "cloud_run"]
interactive_command = "dev"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"